requirements.txt
email_validator==2.0.0.post2
Flask==2.3.2
flask_sqlalchemy==3.0.3
Flask_WTF==1.1.1
mysql_connector_repackaged==0.3.1
pandas==1.5.3
plotly==5.14.1
PyMySQL==1.0.2
SQLAlchemy==2.0.12
WTForms==3.0.1
bcrypt==4.0.1
